+++ tsconfig.json
@@ -1,26 +1,15 @@
 {
-    "compilerOptions": {
-      "target": "es5",
-      "lib": ["dom", "esnext"],
-      "skipLibCheck": true,
-      "esModuleInterop": true,
-      "allowSyntheticDefaultImports": true,
-      "strict": true,
-      "forceConsistentCasingInFileNames": true,
-      "module": "esnext",
-      "moduleResolution": "node",
-      "resolveJsonModule": true,
-      "jsx": "react",
-      "sourceMap": true,
-      "types": ["jest", "node"],
-      "allowJs": true,
-      "noEmit": true
-    },
-    "include": [
-      "./main.ts"
-    ],
-    "exclude": [
-      "node_modules"
-    ],
-  }
-  
+  "compilerOptions": {
+    "module": "commonjs",
+    "noImplicitAny": true,
+    "sourceMap": true,
+    "outDir": "dist",
+    "baseUrl": ".",
+    "paths": {
+      "*": ["node_modules/*"]
+    }
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
